initial packaging for debian
This is a squash of following commits:

-----

initial packaging for debian


-----

fixed mistake about init


-----

debian: fix build dependencies

-----

Makefile: Link to glib-2.0 before libgbinder

Should fix 'DSO missing from command line' errors when linking.

-----

rebased on actual master state:
  - https://github.com/mer-hybris/bluebinder/tree/63b9874fb28516faf6025b986c090d78f79c7bb6

+++ Makefile
@@ -2,7 +2,7 @@
 
 USE_SYSTEMD ?= 1
 
-DEPEND_LIBS = libgbinder glib-2.0
+DEPEND_LIBS = glib-2.0 libgbinder
 ifeq ($(USE_SYSTEMD),1)
 DEPEND_LIBS += libsystemd
 endif
@@ -17,5 +17,5 @@ install:
 	cp bluebinder $(DESTDIR)/usr/sbin
 
 clean:
-	rm bluebinder
+	if test -a "bluebinder"; then rm bluebinder; fi;
